pipeline/requirements.txt
# Core dependencies
requests>=2.31.0
tqdm>=4.66.0
numpy>=1.24.0

# XML parsing
lxml>=4.9.0

# Vector similarity search
faiss-cpu>=1.7.4

# Database connectivity (for loaders in dbtools/)
psycopg2-binary>=2.9.9

# JSON schema validation (optional)
jsonschema>=4.20.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality
ruff>=0.1.0
black>=23.12.0

# Type hints
mypy>=1.7.0
types-requests>=2.31.0
